+++ NEWS.md
@@ -4,6 +4,8 @@ quanteda 0.9.8
 
 ## New Features
 
+*  Added tokens methods for `collocations()` and `kwic()` (0.9.9.11).
+*  Added fast `selectFeatures.tokens()` (0.9.9.11).
 *  Improved `ngrams()` and `joinTokens()` performance for hashed `tokens` class objects (0.9.9.10).
 *  Improved `dfm.character()` by using new `tokens()` constructor to create hashed tokenized texts by default when creating a dfm, resulting in performance gains when constructing a dfm.  Creating a dfm from a hashed `tokens` object is now 4-5 times faster than the older `tokenizedTexts` object (0.9.9.9).
 *  Added new (hashed) `tokens` class object replacing the `tokenizedTextsHashed` class (0.9.9.9).